root/shopping_app/requirement.txt
flask
Flask-SQLAlchemy
pandas
numpy
